# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FarmShield AI",
    "description": "\n**Hyperlocal climate risk infrastructure for smallholder farmers in Kenya.**\n\nFarmShield turns weather-station readings (JKUAT Conduit station) + crop + growth stage + location into:\n\n* a **Farm Risk Score** - drought, flood, heat-stress and crop-health sub-scores (LOW / MEDIUM / HIGH) and an overall 0-100 score, each backed by human-readable *reasons*;\n* **stage-aware advice** for the farmer in English and Kiswahili, delivered by SMS;\n* **parametric insurance trigger signals** insurers and SACCOs can act on.\n\nScoring is deterministic and rule-based (FAO-56 crop water requirements, KALRO crop guides) so every number is auditable.\nPartners authenticate with an `X-API-Key` header on the `/api/v1` routes.\n",
    "contact": {
      "name": "FarmShield AI - Hack the Weather 2026",
      "url": "https://github.com/Millie-source/FarmShield-AI"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.2.0"
  },
  "paths": {
    "/farms": {
      "get": {
        "tags": [
          "farms"
        ],
        "summary": "List farms with their latest risk summary",
        "operationId": "list_farms_farms_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FarmOut"
                  },
                  "type": "array",
                  "title": "Response List Farms Farms Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "farms"
        ],
        "summary": "Register a farm",
        "description": "Registers a farmer (matched by phone number) and a farm plot, then runs a first risk assessment immediately so the dashboard shows a score straight away.",
        "operationId": "create_farm_farms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FarmCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation error (unknown crop, bad phone, out-of-bounds location)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/farms/{farm_id}": {
      "get": {
        "tags": [
          "farms"
        ],
        "summary": "Get one farm",
        "operationId": "get_farm_farms__farm_id__get",
        "parameters": [
          {
            "name": "farm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Farm Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "farms"
        ],
        "summary": "Delete a farm",
        "operationId": "delete_farm_farms__farm_id__delete",
        "parameters": [
          {
            "name": "farm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Farm Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/farms/{farm_id}/weather": {
      "get": {
        "tags": [
          "farms"
        ],
        "summary": "Daily weather history used for scoring",
        "description": "Readings from the active provider (JKUAT Conduit station or the mock replay). Pass `scenario` to preview a mock scenario without switching it globally.",
        "operationId": "farm_weather_farms__farm_id__weather_get",
        "parameters": [
          {
            "name": "farm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Farm Id"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "minimum": 1,
              "description": "Number of days of history",
              "default": 30,
              "title": "Days"
            },
            "description": "Number of days of history"
          },
          {
            "name": "scenario",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "normal",
                    "dry_spell",
                    "heavy_rain"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Force a mock scenario for this call",
              "title": "Scenario"
            },
            "description": "Force a mock scenario for this call"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherHistoryOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/farms/{farm_id}/alerts/preview": {
      "post": {
        "tags": [
          "alerts"
        ],
        "summary": "Preview the SMS that would be sent",
        "description": "Builds the exact <=160-char message from the latest assessment (running one if needed) and applies the alert policy: no SMS for LOW risk unless an insurance trigger fired, no repeat within `ALERT_DEDUPE_HOURS` unless the risk level changed. Nothing is sent or stored.",
        "operationId": "preview_alert_farms__farm_id__alerts_preview_post",
        "parameters": [
          {
            "name": "farm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Farm Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AlertRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "Optional overrides; send `{}` or nothing for defaults",
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertPreviewOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/farms/{farm_id}/alerts/send": {
      "post": {
        "tags": [
          "alerts"
        ],
        "summary": "Send an SMS alert to the farmer",
        "description": "Applies the same policy as `/preview`; when warranted (or `force=true`) sends via the configured sender (Africa's Talking sandbox or console) and records the alert. If the gateway fails the message is logged to the console and the alert is stored with `source: \"fallback\"` - the call still returns 200. A suppressed alert also returns 200 with `sent: false` and the reason.",
        "operationId": "send_alert_farms__farm_id__alerts_send_post",
        "parameters": [
          {
            "name": "farm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Farm Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AlertRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "Optional overrides; send `{}` or nothing for defaults",
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertSendOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/farms/{farm_id}/alerts": {
      "get": {
        "tags": [
          "alerts"
        ],
        "summary": "Alert history for a farm (newest first)",
        "operationId": "list_alerts_farms__farm_id__alerts_get",
        "parameters": [
          {
            "name": "farm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Farm Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlertOut"
                  },
                  "title": "Response List Alerts Farms  Farm Id  Alerts Get"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/farms": {
      "get": {
        "tags": [
          "partners"
        ],
        "summary": "Portfolio: all registered farms with latest risk summary",
        "operationId": "portfolio_api_v1_farms_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FarmOut"
                  },
                  "type": "array",
                  "title": "Response Portfolio Api V1 Farms Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid X-API-Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "PartnerApiKey": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlertPreviewOut": {
        "properties": {
          "farm_id": {
            "type": "integer",
            "title": "Farm Id"
          },
          "assessment_id": {
            "type": "integer",
            "title": "Assessment Id"
          },
          "would_send": {
            "type": "boolean",
            "title": "Would Send"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "description": "forced | first_alert | level_changed:A->B | insurance_trigger | repeat_after_window | duplicate_within_window:Nh | below_alert_threshold:L<M"
          },
          "recipient": {
            "type": "string",
            "title": "Recipient"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Exact SMS text that would be sent"
          },
          "chars": {
            "type": "integer",
            "title": "Chars",
            "description": "Message length (target <= 160 for one SMS segment)"
          },
          "level": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ],
            "title": "Level"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "last_alert_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Alert Id"
          },
          "last_alert_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Alert At"
          },
          "sender": {
            "type": "string",
            "title": "Sender",
            "description": "Configured sender: console | africastalking"
          }
        },
        "type": "object",
        "required": [
          "farm_id",
          "assessment_id",
          "would_send",
          "reason",
          "recipient",
          "language",
          "message",
          "chars",
          "level",
          "score",
          "sender"
        ],
        "title": "AlertPreviewOut",
        "description": "What `POST /farms/{id}/alerts/send` would do right now, without sending.",
        "example": {
          "assessment_id": 12,
          "chars": 139,
          "farm_id": 1,
          "language": "sw",
          "last_alert_at": "2026-09-03T06:00:00Z",
          "last_alert_id": 3,
          "level": "HIGH",
          "message": "FarmShield: Kamau Maize Plot hatari JUU (72/100). Mahindi kutoa maua. Mwagilia ndani ya saa 24-48. Usiweke mbolea. Bima inaweza kulipa.",
          "reason": "level_changed:MEDIUM->HIGH",
          "recipient": "+254711000001",
          "score": 72,
          "sender": "console",
          "would_send": true
        }
      },
      "AlertRequest": {
        "properties": {
          "language": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "en",
                  "sw"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Language",
            "description": "Override the farmer's preferred language"
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "description": "Send even if the dedupe policy says the alert is not warranted (dashboard button)",
            "default": false
          },
          "message": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 320
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Custom text instead of the generated SMS (send only; trimmed to 160 chars)"
          }
        },
        "type": "object",
        "title": "AlertRequest",
        "description": "Options for previewing / sending an SMS alert. All fields optional.",
        "example": {
          "force": false,
          "language": "sw"
        }
      },
      "AlertSendOut": {
        "properties": {
          "farm_id": {
            "type": "integer",
            "title": "Farm Id"
          },
          "sent": {
            "type": "boolean",
            "title": "Sent",
            "description": "False when the dedupe policy suppressed the SMS (HTTP still 200)"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "alert": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AlertOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "The recorded alert when sent"
          }
        },
        "type": "object",
        "required": [
          "farm_id",
          "sent",
          "reason"
        ],
        "title": "AlertSendOut",
        "example": {
          "farm_id": 1,
          "reason": "duplicate_within_window:6h",
          "sent": false
        }
      },
      "ApiError": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ApiError",
        "example": {
          "detail": "Farm 42 not found"
        }
      },
...
      "FarmCreate": {
        "properties": {
          "farmer_name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Farmer Name",
            "description": "Farmer's full name"
          },
          "phone": {
            "type": "string",
            "title": "Phone",
            "description": "Mobile number in E.164 format (+2547...). Used for SMS alerts."
          },
          "language": {
            "type": "string",
            "enum": [
              "en",
              "sw"
            ],
            "title": "Language",
            "description": "Preferred language for advice and SMS",
            "default": "en"
          },
          "farm_name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Farm Name",
            "description": "Human-friendly farm / plot name"
          },
          "crop": {
            "type": "string",
            "title": "Crop",
            "description": "One of: maize, beans, potatoes, tomatoes, kale"
          },
          "planting_date": {
            "type": "string",
            "format": "date",
            "title": "Planting Date",
            "description": "Date the crop was planted / transplanted"
          },
          "lat": {
            "type": "number",
            "maximum": 5.5,
            "minimum": -5.5,
            "title": "Lat",
            "description": "Latitude (Kenya bounds)"
          },
          "lon": {
            "type": "number",
            "maximum": 42.5,
            "minimum": 33.5,
            "title": "Lon",
            "description": "Longitude (Kenya bounds)"
          },
          "area_ha": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Ha",
            "description": "Plot size in hectares (optional)"
          }
        },
        "type": "object",
        "required": [
          "farmer_name",
          "phone",
          "farm_name",
          "crop",
          "planting_date",
          "lat",
          "lon"
        ],
        "title": "FarmCreate",
        "example": {
          "area_ha": 0.8,
          "crop": "maize",
          "farm_name": "Kamau Maize Plot",
          "farmer_name": "Wanjiku Kamau",
          "language": "sw",
          "lat": -1.098,
          "lon": 37.012,
          "phone": "+254711000001",
          "planting_date": "2026-07-03"
        }
      },
      "FarmOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "farm_name": {
            "type": "string",
            "title": "Farm Name"
          },
          "farmer_name": {
            "type": "string",
            "title": "Farmer Name"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "crop": {
            "type": "string",
            "title": "Crop"
          },
          "crop_display": {
            "type": "string",
            "title": "Crop Display"
          },
          "planting_date": {
            "type": "string",
            "format": "date",
            "title": "Planting Date"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "title": "Lon"
          },
          "area_ha": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Ha"
          },
          "county": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "County"
          },
          "stage": {
            "type": "string",
            "title": "Stage",
            "description": "Current growth stage derived from planting date"
          },
          "days_after_planting": {
            "type": "integer",
            "title": "Days After Planting"
          },
          "latest_risk": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "farm_name",
          "farmer_name",
          "phone",
          "language",
          "crop",
          "crop_display",
          "planting_date",
          "lat",
          "lon",
          "area_ha",
          "county",
          "stage",
          "days_after_planting"
        ],
        "title": "FarmOut",
        "example": {
          "area_ha": 0.8,
          "county": "Kiambu",
          "crop": "maize",
          "crop_display": "Maize",
          "days_after_planting": 62,
          "farm_name": "Kamau Maize Plot",
          "farmer_name": "Wanjiku Kamau",
          "id": 1,
          "language": "sw",
          "lat": -1.098,
          "latest_risk": {
            "assessed_at": "2026-09-03T08:00:00Z",
            "assessment_id": 12,
            "farm_id": 1,
            "insurance_triggered": true,
            "overall_label": "HIGH CLIMATE RISK",
            "overall_level": "HIGH",
            "overall_score": 72,
            "scenario": "dry_spell",
            "stage": "flowering"
          },
          "lon": 37.012,
          "phone": "+254711000001",
          "planting_date": "2026-07-03",
          "stage": "flowering"
        }
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "WeatherHistoryOut": {
        "properties": {
          "farm_id": {
            "type": "integer",
            "title": "Farm Id"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Data source id, e.g. mock:dry_spell or conduit:jkuat"
          },
          "days": {
            "type": "integer",
            "title": "Days"
          },
          "readings": {
            "items": {
              "$ref": "#/components/schemas/WeatherReadingOut"
            },
            "type": "array",
            "title": "Readings"
          }
        },
        "type": "object",
        "required": [
          "farm_id",
          "source",
          "days",
          "readings"
        ],
        "title": "WeatherHistoryOut"
      },
...
    },
    "securitySchemes": {
      "PartnerApiKey": {
        "type": "apiKey",
        "description": "Partner API key issued to insurers, banks, SACCOs and agribusinesses. Demo keys: `fs_demo_acme_insurance_2026` (acme-insurance), `fs_demo_harvest_sacco_2026` (harvest-sacco).",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}